Moved repository info to separate reducer
